Accomplish User Story 5 with API endpoint for user role remote updating
Authorization was possible with application passwords activated.
However, the option was not showing because the website does not use HTTPS.
Finally, a filter function helped show the option on WP DB.

Before, tried migrating the website to HTTPS - it was a catastrophe.
A lot of changes was required and not enough time.
Point noted to start project in TSL mode before creating a WP website.